rip out forms from HomeView component
